Use constant-time comparison for tags.
This is as close to constant-time as I can figure out how to make this.

I originally intended to use a boolean instead of a counter, but there
is no non-short-circuiting booelan `and` in Elixir.

I also considered adding `0` instead of `2` but thought that might get
optimized away somewhere, too.