Simplify `process_message` a bit.
This is a bit less imperative. No sense chopping up the message several
times if we can get the BEAM to do most of the heavy-lifting.

The final message clause is a bit "meh".. but it works.

The piping there precludes a tail-call back to the empty string clause
without some shenanigans. I don't see how it could too much impact
optimizations.  I suppose if something further happens in the empty case
We can break up this line and turn it into two with the call.